# src/main/java/com/aforo/apigee/dto/response/DeveloperResponse.java
package com.aforo.apigee.dto.response;

import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

import java.util.Map;

@Data
@Builder
@NoArgsConstructor
@AllArgsConstructor
public class DeveloperResponse {
    private String id;
    private String email;
    private Map<String, String> attributes;
}